# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/calls": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calls"
                ],
                "summary": "Начать звонок",
                "parameters": [
                    {
                        "description": "чат и тип звонка",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modules_calls_handler.startCallRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/calls/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calls"
                ],
                "summary": "История звонков",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "лимит",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/calls/ws": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calls"
                ],
                "summary": "WebRTC-сигналинг (WebSocket)",
                "responses": {
                    "101": {
                        "description": "Switching Protocols"
                    }
                }
            }
        },
        "/calls/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calls"
                ],
                "summary": "Получить звонок по id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id звонка",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/calls/{id}/end": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calls"
                ],
                "summary": "Завершить звонок",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id звонка",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_jeogram_messenger_internal_pkg_response.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_jeogram_messenger_internal_pkg_response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.APIError"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_jeogram_messenger_internal_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_jeogram_messenger_internal_pkg_response.Meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "modules_calls_handler.startCallRequest": {
            "type": "object",
            "required": [
                "chat_id",
                "type"
            ],
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "audio",
                        "video"
                    ]
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}